+++ ProgramFunctions/Substitutions.fnc.php
@@ -34,8 +34,7 @@ function SubstitutionsInput( $substitutions )
 
 	$id++;
 
-	// @link https://stackoverflow.com/questions/10672586/how-to-make-select-elements-shrink-to-max-width-percent-style-within-fieldset
-	$input_html = '<div style="display: inline-block; max-width: 440px;">' . SelectInput(
+	$input_html = SelectInput(
 		'',
 		'substitutions_input_' . $id,
 		'',
@@ -43,7 +42,7 @@ function SubstitutionsInput( $substitutions )
 		$allow_na,
 		'autocomplete="off"',
 		$div
-	) . '</div>';
+	);
 
 	$code_value = key( $substitutions );