+++ modules/Discipline/StudentFieldBreakdown.php
@@ -262,9 +262,7 @@ if ( ! $_REQUEST['modfunc'] )
 	$categories_RET = DBGet( "SELECT ID,TITLE
 		FROM STUDENT_FIELD_CATEGORIES", [], [ 'ID' ] );
 
-	// @link https://stackoverflow.com/questions/10672586/how-to-make-select-elements-shrink-to-max-width-percent-style-within-fieldset
-	$select = '<div style="display: inline-block; max-width: 440px;">
-		<select name="field_id" onchange="ajaxPostForm(this.form,true);">';
+	$select = '<select name="field_id" onchange="ajaxPostForm(this.form,true);">';
 
 	$select .= '<option value="">' . _( 'Please choose a student field' ) . '</option>';
 
@@ -300,7 +298,7 @@ if ( ! $_REQUEST['modfunc'] )
 	}
 
 	$select .= '</select>
-		<label for="field_id" class="a11y-hidden">' . _( 'Category' ) . '</label></div>';
+		<label for="field_id" class="a11y-hidden">' . _( 'Category' ) . '</label>';
 
 	$advanced_link = ' <a href="' . PreparePHP_SELF( $_REQUEST, [ 'search_modfunc' ], [
 		'modfunc' => 'search',